Separate Enum From Other Class
