+++ src/main/java/flab/delideli/annotation/LoginUserLevel.java
@@ -1,5 +1,6 @@
 package flab.delideli.annotation;
 
+import flab.delideli.enums.UserLevel;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -10,11 +11,5 @@ import java.lang.annotation.Target;
 public @interface LoginUserLevel {
 
 	UserLevel role();
-
-	enum UserLevel {
-		ADMIN_LEVEL,
-		OWNER_LEVEL,
-		MEMBER_LEVEL
-	}
 
 }
