+++ src/main/java/flab/delideli/dto/MemberDTO.java
@@ -1,6 +1,5 @@
 package flab.delideli.dto;
 
-import flab.delideli.annotation.LoginUserLevel.UserLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,8 +9,6 @@ import javax.validation.constraints.NotBlank;
 @Getter
 @NoArgsConstructor
 public class MemberDTO {
-
-    private static final UserLevel userLevel = UserLevel.MEMBER_LEVEL;
 
     @NotBlank
     private String userid;
