+++ src/main/java/flab/delideli/dto/OwnerDTO.java
@@ -1,6 +1,5 @@
 package flab.delideli.dto;
 
-import flab.delideli.annotation.LoginUserLevel.UserLevel;
 import javax.validation.constraints.NotBlank;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,8 +8,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class OwnerDTO {
-
-	private static final UserLevel userLevel = UserLevel.OWNER_LEVEL;
 
 	@NotBlank
 	private String ownerId;
