+++ src/main/java/flab/delideli/enums/UserLevel.java
@@ -1,0 +1,9 @@
+package flab.delideli.enums;
+
+public enum UserLevel {
+
+		ADMIN_LEVEL,
+		OWNER_LEVEL,
+		MEMBER_LEVEL
+
+}
