+++ src/main/java/flab/delideli/interceptor/LoginInterceptor.java
@@ -1,11 +1,10 @@
 package flab.delideli.interceptor;
 
 import flab.delideli.annotation.LoginUserLevel;
-import flab.delideli.dto.LoginDTO;
-import flab.delideli.dto.OwnerDTO;
+//import flab.delideli.dto.LoginDTO;
+//import flab.delideli.enums.UserLevel;
 import flab.delideli.exception.UnauthorizedException;
 import flab.delideli.service.LoginService;
-import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
@@ -38,19 +37,6 @@ public class LoginInterceptor implements HandlerInterceptor {
         LoginUserLevel loginUserLevel = handlerMethod.getMethodAnnotation(LoginUserLevel.class);
         if (loginUserLevel == null) {
             return true;
-        }
-
-        String currentUserId = SessionloginService.getSessionUserId(request);
-        String role = loginUserLevel.role().toString();
-        LoginDTO loginDTO = (LoginDTO) request.getSession().getAttribute("USER_ID");
-        LoginUserLevel userLevel = loginDTO.getLoginUserLevel();
-        if (currentUserId != null) {
-            if ("OWNER_LEVEL".equals(role) && "OWNER_LEVEL".equals(userLevel)) {
-                return true;
-            }
-
-            return true;
-
         }
 
         throw new UnauthorizedException();
