+++ src/main/java/flab/delideli/service/LoginService.java
@@ -4,6 +4,6 @@ import javax.servlet.http.HttpServletRequest;
 
 public interface LoginService {
 
-    public String getSessionUserId(HttpServletRequest request);
+    String getSessionUserId(HttpServletRequest request);
 
-    }
+}
